Move fakesocket package
No functional changes
