+++ pkg/fakesocket/fake.go
